Coparticipação

+++ application/views/emergencia/solicitacoes-paciente.php
@@ -89,7 +89,7 @@
                     <? if ($perfil_id == 1) { ?>
 
 
-                        <td width="100px;"><div class="bt_linkm"><a href="<?= base_url() ?>ambulatorio/exametemp/excluirpaciente/<?= $paciente_id ?>">Excluir</a></div></td>
+                        <td width="100px;"><div class="bt_linkm"><a onclick="javascript: return confirm('Deseja realmente excluir o cliente?');" href="<?= base_url() ?>ambulatorio/exametemp/excluirpaciente/<?= $paciente_id ?>">Excluir</a></div></td>
                     <? }
                     ?>
                 </tr>
